# Master imp add formula vectorization laa