--- a/src/functions/helper_lookup.ts
+++ b/src/functions/helper_lookup.ts
@@ -1,6 +1,6 @@
 import { toZone, zoneToXc } from "../helpers";
 import { _t } from "../translation";
-import { CellPosition, EvalContext, FPayload, Getters, Maybe, UID } from "../types";
+import { CellPosition, EvalContext, FunctionResultObject, Getters, Maybe, UID } from "../types";
 import { EvaluationError, InvalidReferenceError } from "../types/errors";
 import { PivotCoreDefinition } from "../types/pivot";
 
@@ -29,7 +29,7 @@ export function assertMeasureExist(pivotId: UID, measure: string, getters: Gette
   }
 }
 
-export function assertDomainLength(domain: Maybe<FPayload>[]) {
+export function assertDomainLength(domain: Maybe<FunctionResultObject>[]) {
   if (domain.length % 2 !== 0) {
     throw new EvaluationError(_t("Function PIVOT takes an even number of arguments."));
   }