… src/functions/module_array.ts
… src/functions/module_database.ts
… src/functions/module_date.ts
… src/functions/module_filter.ts
… src/functions/module_financial.ts
… src/functions/module_info.ts
… src/functions/module_lookup.ts
… src/functions/module_math.ts
… src/functions/module_operators.ts
… src/functions/module_statistical.ts
… src/functions/module_text.ts
… src/helpers/cells/cell_evaluation.ts
… src/helpers/format.ts
… src/helpers/pivot/pivot_helpers.ts
… src/helpers/pivot/spreadsheet_pivot/spreadsheet_pivot.ts
… src/plugins/core/merge.ts
… src/plugins/ui_core_views/cell_evaluation/compilation_parameters.ts
… src/plugins/ui_core_views/cell_evaluation/evaluator.ts
… src/plugins/ui_core_views/cell_evaluation/formula_dependency_graph.ts
… src/plugins/ui_core_views/pivot_ui.ts
… src/types/cells.ts
… src/types/functions.ts
… src/types/misc.ts
… src/types/pivot.ts
… src/types/pivot_runtime.ts
… tests/evaluation/evaluation.test.ts
… tests/evaluation/evaluation_formula_array.test.ts
… tests/functions/functions.test.ts
… tests/functions/module_logical.test.ts
… tests/functions/vectorization.test.ts