## Concerning performances:

I don't measure any significant change for the formulas dataset and for the array formulas dataset.

Regarding the addition of vectorized formulas, a large dataset was created too.
It has exactly the same data as in the formulas data set. See the difference above : 

| makeLargeDataset                    | Evaluate All Cells | Model Created In |
|---------------------|--------------------|------------------|
| **formulas**        | ~ 270 ms                | ~ 2000 ms             |
| **vectorizedFormulas** | ~ 350 ms                | ~ 380 ms              |




Task: : [3356482](https://www.odoo.com/web#id=3356482&action=333&active_id=2328&model=project.task&view_type=form&cids=1&menu_id=4720)


- [ ] feature is organized in plugin, or UI components
- [ ] support of duplicate sheet (deep copy)
- [ ] in model/core: ranges are Range object, and can be adapted (adaptRanges)
- [ ] in model/UI: ranges are strings (to show the user)
- [ ] undo-able commands (uses this.history.update)
- [ ] multiuser-able commands (has inverse commands and transformations where needed)
- [ ] new/updated/removed commands are documented
- [ ] exportable in excel
- [ ] translations (\_t("qmsdf %s", abc))
- [ ] unit tested
- [ ] clean commented code
- [ ] track breaking changes
- [ ] doc is rebuild (npm run doc)
- [ ] status is correct in Odoo